requirements.txt
yfinance==0.2.38
requests==2.31.0
google-api-python-client==2.126.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
apscheduler==3.10.4
python-dotenv==1.0.1
flask==3.0.3
google-generativeai==0.5.2
pytelegrambotapi==4.17.0
pandas==2.2.2
gunicorn==22.0.0
moviepy==1.0.3
numpy==1.26.4
